refactor: move BNCollection for his own file
